+++ Project/hardware/inc/gyroCompass.h
@@ -14,11 +14,9 @@
 // Function prototypes
 uint8_t magnetometer_init(void);
 uint8_t magnetometer_exit(void);
-uint8_t magnetometer_get_heading(double *aHeading);
+uint8_t gyroCompass_get_heading(double *aHeading);
 
 uint8_t gyroscope_init(void);
 uint8_t gyroscope_exit(void);
-uint8_t gyroscope_get_acceleration(double *aAcceleration);
-uint8_t gyroscope_get_gyro(double *aGyro);
 
 #endif /* GYROCOMPASS_H */
