--- Project/hardware/inc/headingCalculator.h
@@ -1,4 +1,0 @@
-#ifndef HEADINGCALCULATOR_H
-#define HEADINGCALCULATOR_H
-
-#endif /* HEADINGCALCULATOR_H */
